lagde ny komponent RedaksjoneltInnhold for at lenker skal åpnes i nye faner. trakk også ut types som ble brukt av oppskrift
